+++ src/reuse/init.py
@@ -7,14 +7,13 @@
 import sys
 from gettext import gettext as _
 from inspect import cleandoc
-from itertools import chain
 from pathlib import Path
 from typing import List
 
 import requests
 
-from ._licenses import EXCEPTION_MAP, LICENSE_MAP
-from ._util import PathType
+from ._licenses import ALL_NON_DEPRECATED_MAP
+from ._util import PathType, print_incorrect_spdx_identifier
 from .download import _path_to_license_file, put_license_in_file
 from .project import Project
 from .vcs import find_root
@@ -44,19 +43,8 @@ def prompt_licenses(out=sys.stdout) -> List[str]:
         out.write("\n")
         if not result:
             return licenses
-        if result not in chain(LICENSE_MAP, EXCEPTION_MAP):
-            out.write(
-                _("'{}' is not a valid SPDX License Identifier.").format(
-                    result
-                )
-            )
-            out.write("\n")
-            out.write(
-                _(
-                    "See <https://spdx.org/licenses/> for a list of valid "
-                    "SPDX License Identifiers."
-                )
-            )
+        if result not in ALL_NON_DEPRECATED_MAP:
+            print_incorrect_spdx_identifier(result, out=out)
             out.write("\n\n")
         else:
             licenses.append(result)